Conflicts:
	test/unit/specs/vuex/data/actions.spec.js

+++ src/components/Security/Roles/RoleItem.vue
@@ -18,9 +18,16 @@
     </label>
 
     <div class="right actions">
-      <a href="#" @click.prevent="update"><i class="fa fa-pencil"></i></a>
+      <a href="#" @click.prevent="update"
+          v-title="{active: !canEditRole(), title: 'You are not allowed to edit this role'}">
+          <i class="fa fa-pencil" :class="{'disabled': !canEditRole()}"></i>
+      </a>
       <dropdown :id="document.id" class="icon-black">
-        <li><a @click="deleteDocument(document.id)">Delete</a></li>
+        <li><a @click="deleteDocument(document.id)"
+               :class="{'disabled': !canDeleteRole()}"
+               v-title="{active: !canDeleteRole(), title: 'You are not allowed to delete this role'}">
+          Delete</a>
+        </li>
       </dropdown>
     </div>
 
@@ -33,6 +40,8 @@
 <script>
 import Dropdown from '../../Materialize/Dropdown'
 import jsonFormatter from '../../../directives/json-formatter.directive'
+import { canEditRole, canDeleteRole } from '../../../services/userAuthorization'
+import title from '../../../directives/title.directive'
 
 export default {
   name: 'RoleItem',
@@ -44,7 +53,8 @@ export default {
     Dropdown
   },
   directives: {
-    jsonFormatter
+    jsonFormatter,
+    title
   },
   data () {
     return {
@@ -59,11 +69,17 @@ export default {
       this.$dispatch('checkbox-click', this.document.id)
     },
     deleteDocument () {
-      this.$dispatch('delete-document', this.document.id)
+      if (this.canDeleteRole()) {
+        this.$dispatch('delete-document', this.document.id)
+      }
     },
     update () {
-      this.$router.go({name: 'SecurityRolesUpdate', params: {id: this.document.id}})
-    }
+      if (this.canEditRole()) {
+        this.$router.go({name: 'SecurityRolesUpdate', params: {id: this.document.id}})
+      }
+    },
+    canEditRole,
+    canDeleteRole
   }
 }
 </script>
